MaxCut mapping to QAOA PQC work
(*In Progress*) Creating a general definition that can map a MaxCut graph to a QAOA PQC circuit for a large class of graphs. Also exploring the parameter optimization of a depth 2 QAOA circuit with the Adam optimizer. Will continue to look at different optimizers within TF that can be applied to PQCs through TFQ.